# perf: optimize relational rule's stopBy: neighbor