## Idea

We can skip creating a tree-cursor if we just want to match one node around. Creating a tree-cursor is quite expensive.

## Implementation
`stop_by`'s `find` now gets three arguments: `once`, `multi` and `finder`. All of the three are closures. We create iterator and, thusly, cursor only when `stop_by` is `end`.

## Benchmark 

Before:

```sh
 hyperfine -i 'sg scan -c ../eslint/sgconfig.yml src > /dev/null'
Benchmark 1: sg scan -c ../eslint/sgconfig.yml src > /dev/null
  Time (mean ± σ):      3.310 s ±  0.055 s    [User: 9.473 s, System: 0.128 s]
  Range (min … max):    3.235 s …  3.405 s    10 runs

  Warning: Ignoring non-zero exit code.
```

After:
```shell
hyperfine -i 'sg scan -c ../eslint/sgconfig.yml src > /dev/null'
Benchmark 1: sg scan -c ../eslint/sgconfig.yml src > /dev/null
  Time (mean ± σ):      2.990 s ±  0.069 s    [User: 8.675 s, System: 0.129 s]
  Range (min … max):    2.880 s …  3.087 s    10 runs

  Warning: Ignoring non-zero exit code.
```

I cannot believe it saves yet another 9% wall time and 8% CPU time! 😍 

partially fix #358.